moving rpc test utils to testutil.go
